코드 컴포넌트와 행사기간 시맨틱 태그 변경
